Move SimpleTransfer to org.opentripplanner.model
